requirements.txt
sentence_transformers
pinecone-client[grpc]
pandas
